# src/main/resources/api.yml
openapi: 3.1.0
info:
  version: v1
  title: content-vault API
  description: content-vault RESTFul api for openapi 3.1
  termsOfService: 'https://jumkid.com/terms-of-use'
  contact:
    name: Chooli Yip
    url: jumkid.com
    email: chooli.yip@gmail.com
  license:
    name: Jumkid Public License
    url: 'http://license.jumkid.com'
servers:
  - description: Testing
    url: 'http://test.jumkid.com'
  - description: Live production
    url: 'http://api.jumkid.com'
...
security:
  - bearerAuth: []
paths:
...
  '/file/upload':
    post:
      tags:
        - v1
      operationId: upload
      description: file upload
      parameters:
        - name: title
          in: query
          required: false
          description: title of file
          schema:
            type: string
        - name: content
          in: query
          required: false
          description: content of file
          schema:
            type: string
        - name: tags
          in: query
          required: false
          description: tags of file
          schema:
            type: array  
            format: string
        - name: accessScope
          in: query
          required: true
          schema:
            $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/AccessScope     

      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                tags:
                  type: array
                  format: string
                  
      responses:
        '202':
          description: upload a new content in file format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaFile'
  /file/multipleUpload:
    post:
      tags:
        - v1
      operationId: multipleUpload
      description: multiple files upload
      parameters:
        - name: tags
          in: query
          required: false
          description: tags of files
          schema:
            type: array
            format: string

      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  format: binary
                tags:
                  type: array
                  format: string

      responses:
        '202':
          description: upload multiple content in file format
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MediaFile'
  '/file/download/{id}':
    get:
      tags:
        - v1
      operationId: download
      description: download file with the identifier
      parameters:
        - name: id
          in: path
          required: true
          description: mediafile identifier
          schema:
            type: string
      responses:
        '200':
          description: binary file as response
...
components:
  schemas:
    MediaFile:
      description: Model of media file
      properties:
        id:
          type: string
          example: 1mWG-nQBWmmN0po1BUoG
        filename:
          type: string
          example: foo.txt
        mimeType:
          type: string
          example: text/plain
        size:
          type: integer
          description: file size in byte
        title:
          type: string
        content:
          type: string
        activated:
          type: boolean
        accessScope:
          $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/AccessScope
        createdBy:
          type: string
        creationDate:
          $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/LocalDateTime
        modifiedBy:
          type: string
        modifiedOn:
          $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/LocalDateTime
        props:
          type: array
          items:
            description: list of additional properties
            properties:
              name:
                type: string
              textValue:
                type: string
              dateValue:
                $ref: https://raw.githubusercontent.com/jumkid/share-jar/master/src/main/resources/share.yaml#/components/schemas/LocalDateTime
              numberValue:
                type: integer
        tags:
          type: array
          items:
            description: list of tags
            type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/MediaFile'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT